Code Review of AnalyseText method
list of improvements:

-Duplicate For each loops of character analysis combined into a single loop compacting code.

-Unnecessary temporary count variables removed: 
Using the values.DefaultIfEmpty<int>(0) method all the empty values of initialized "values" array become 0.
Instead of incrementing temporary variables vowelCount, consonantCount, etc. the designated index of "values" is incremented directly.

-Non consonants being counted as consonants:
Even though you were excluding vowels as part of the consonant count, other characters such as punctuation, spaces etc. were being counted as consonants
Instead of a simple else statement I used a else if with the condition that the character was a letter (using Char.IsLetter(c) ) but not a vowel.

Apart from that this I like everything else about it :)